LayoutTests/http/tests/site-isolation/history/iframe-redirect-back-expected.txt

Verifies that a cross-origin iframe that was navigated via client-side redirect is correctly restored after back navigation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS iframeUrl.includes('id=redirected') is true
PASS iframeUrl.includes('redirect-to') is false
PASS successfullyParsed is true

TEST COMPLETE
